requirements.txt
-i https://pypi.org/simple
aiogram==2.3
aiohttp-socks==0.2.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
babel==2.7.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
furl==2.1.0
idna==2.8
multidict==4.5.2
orderedmultidict==1.0.1
python-box==3.4.5
pytz==2019.3
six==1.12.0
soupsieve==1.9.4
tinydb==3.15.0
yarl==1.3.0
